requirements.txt
bidict==0.22.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
colorlog==6.6.0
cryptography==37.0.4
Cython==0.29.32
dearpygui==1.6.2
dnspython==2.2.1
eventlet==0.33.1
greenlet==1.1.3
idna==3.3
lxml==4.9.1
numpy==1.23.2
opcua==0.98.13
opencv-python==4.5.3.56
pycparser==2.21
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.2.1
pyzmq==23.2.1
requests==2.28.1
simplejpeg==1.6.5
six==1.16.0
tqdm==4.64.0
urllib3==1.26.12
vidgear==0.2.6
websocket-client==1.4.0
zmq==0.0.0
